continue restructuring
- Dependeny
- Milestone
- OtherPreRelease
- PreRelease
- ReleaseCandidate
- ScalaTargets
- SemanticVersion
