+++ data/src/test/scala/ch.epfl.scala.index/cleanup/ArtifactNameParserTests.scala
@@ -2,7 +2,7 @@ package ch.epfl.scala.index
 package data
 package cleanup
 
-import ch.epfl.scala.index.model.release.ScalaTargets
+import ch.epfl.scala.index.model.release.{ReleaseCandidate, ScalaTargets, SemanticVersion}
 import model._
 import utest._
