+++ data/src/test/scala/ch.epfl.scala.index/cleanup/SemanticVersionTest.scala
@@ -2,8 +2,8 @@ package ch.epfl.scala.index
 package data
 package cleanup
 
+import ch.epfl.scala.index.model.release.{Milestone, OtherPreRelease, ReleaseCandidate, SemanticVersion}
 import model._
-
 import utest._
 
 object SemanticVersionTest extends TestSuite{
