+++ model/src/main/scala/ch.epfl.scala.index.model/release/Dependency.scala
@@ -1,5 +1,4 @@
 package ch.epfl.scala.index.model.release
-
 import ch.epfl.scala.index.model.misc.GeneralReference
 
 /**
