+++ model/src/main/scala/ch.epfl.scala.index.model/release/Milestone.scala
@@ -1,0 +1,7 @@
+package ch.epfl.scala.index.model.release
+
+/**
+ * Milestone
+ * @param m the milestone number
+ */
+case class Milestone(m: Long) extends PreRelease
