+++ model/src/main/scala/ch.epfl.scala.index.model/release/OtherPreRelease.scala
@@ -1,0 +1,7 @@
+package ch.epfl.scala.index.model.release
+
+/**
+ * Some other pre release
+ * @param o the pre release name
+ */
+case class OtherPreRelease(o: String) extends PreRelease
