+++ model/src/main/scala/ch.epfl.scala.index.model/release/PreRelease.scala
@@ -1,0 +1,6 @@
+package ch.epfl.scala.index.model.release
+
+/**
+ * Pre release trait
+ */
+trait PreRelease
