+++ template/src/main/twirl/ch.epfl.scala.index.views/artifact.scala.html
@@ -1,9 +1,8 @@
 @import ch.epfl.scala.index.model.Project
 @import ch.epfl.scala.index.model.Artifact
-@import ch.epfl.scala.index.model.SemanticVersion
 @import ch.epfl.scala.index.model.Release
 @import ch.epfl.scala.index.model.misc.UserInfo
-@import ch.epfl.scala.index.model.release.{Dependency, JavaDependency, ScalaDependency}
+@import ch.epfl.scala.index.model.release.{Dependency, JavaDependency, ScalaDependency, SemanticVersion}
 @(project: Project, artifact: Artifact.Reference, version: Option[SemanticVersion], latestStableReleaseOrSelect: List[Release], user: Option[UserInfo])
 
 @staticJs = {
