+++ template/src/main/twirl/ch.epfl.scala.index.views/headproject.scala.html
@@ -1,4 +1,5 @@
 @import ch.epfl.scala.index.model._
+@import ch.epfl.scala.index.model.release.SemanticVersion
 @(project: Project, artifactRef: Artifact.Reference, version: Option[SemanticVersion])
 <div class="head-project">
   <div class="container">
